refactors to increase modularity in backend
